added a new and update js file
